### 2023-11-09 #6048 @editorialbot
[PRE REVIEW]: sectionproperties: A Python package for the analysis of arbitrary cross-sections using the finite element method
**Submitting author:** @robbievanleeuwen (<a href="http://orcid.org/0009-0004-8056-3977">Robbie van Leeuwen</a>)
**Repository:** https://github.com/robbievanleeuwen/section-properties
**Branch with paper.md** (empty if default branch):
**Version:** v3.1.2
