### 2023-11-23 #6048 @prashjha
Howdy @kinnala, would you be interested in reviewing the JOSS submission above ([pdf link](https://github.com/openjournals/joss-papers/blob/joss.06048/joss.06048/10.21105.joss.06048.pdf))? The corresponding software is at this [link](https://github.com/robbievanleeuwen/section-properties).

This [link](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) covers pretty much everything about the review process; I will be happy to help with more information.

Please let me know if you are interested. Thank you!
