### 2023-11-24 #6048 @dbrizard
Hey @prashjha, I just succeeded in installing python 3.9 on my machine and was therefore able to also install `sectionproperties`.
So I will be pleased to review this submission !
