### 2023-11-29 #6048 @valentineap
I’m sorry, while I’d like to help I’m afraid I don’t have time to do this justice at the moment.

Thanks,
Andrew

--
Dr Andrew Valentine
Associate Professor, Department of Earth Sciences, Durham University
Hon. Senior Lecturer, Research School of Earth Sciences, The Australian National University
