### 2023-12-01 8f04f33
Bump virtualenv from 20.24.6 to 20.25.0 in /.github/workflows
Bumps [virtualenv](https://github.com/pypa/virtualenv) from 20.24.6 to 20.25.0.
- [Release notes](https://github.com/pypa/virtualenv/releases)
- [Changelog](https://github.com/pypa/virtualenv/blob/main/docs/changelog.rst)
- [Commits](https://github.com/pypa/virtualenv/compare/20.24.6...20.25.0)

---
updated-dependencies:
- dependency-name: virtualenv
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
