+++ .github/workflows/constraints.txt
@@ -2,4 +2,4 @@ pip==23.3.1
 nox==2023.4.22
 nox-poetry==1.0.3
 poetry==1.6.1
-virtualenv==20.24.6
+virtualenv==20.25.0
