### 2023-12-19 09f41e4
Bump poetry from 1.6.1 to 1.7.1 in /.github/workflows
Bumps [poetry](https://github.com/python-poetry/poetry) from 1.6.1 to 1.7.1.
- [Release notes](https://github.com/python-poetry/poetry/releases)
- [Changelog](https://github.com/python-poetry/poetry/blob/master/CHANGELOG.md)
- [Commits](https://github.com/python-poetry/poetry/compare/1.6.1...1.7.1)

---
updated-dependencies:
- dependency-name: poetry
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
