+++ pyproject.toml
@@ -65,7 +65,7 @@ intel-openmp = { version = "==2023.2.0", optional = true}
 
 [tool.poetry.group.dev.dependencies]
 black = "^23.10.1"
-coverage = { extras = ["toml"], version = "^7.3.2" }
+coverage = { extras = ["toml"], version = "^7.3.3" }
 darglint = "^1.8.1"
 flake8 = "^6.1.0"
 flake8-bugbear = "^23.9.16"
