### 2023-12-19 565ceeb
Bump flake8-bugbear from 23.9.16 to 23.12.2
