+++ pyproject.toml
@@ -68,7 +68,7 @@ black = "^23.10.1"
 coverage = { extras = ["toml"], version = "^7.3.2" }
 darglint = "^1.8.1"
 flake8 = "^6.1.0"
-flake8-bugbear = "^23.9.16"
+flake8-bugbear = "^23.12.2"
 flake8-docstrings = "^1.7.0"
 flake8-pytest-style = "^1.7.2"
 flake8-rst-docstrings = "^0.3.0"
