### 2023-12-19 c62e1bb
Bump ipython from 8.17.2 to 8.18.1
