+++ docs/requirements.txt
@@ -1,6 +1,6 @@
 furo==2023.9.10
 ipykernel==6.26.0
-ipython==8.17.2
+ipython==8.18.1
 ipywidgets==8.1.1
 nbconvert==7.10.0
 nbsphinx==0.9.3
