+++ pyproject.toml
@@ -74,7 +74,7 @@ flake8-pytest-style = "^1.7.2"
 flake8-rst-docstrings = "^0.3.0"
 furo = "^2023.9.10"
 ipykernel = "^6.26.0"
-ipython = "^8.17.2"
+ipython = "^8.18.1"
 ipywidgets = "^8.1.1"
 isort = "^5.12.0"
 mypy = "^1.6.1"
