+++ pyproject.toml
@@ -76,7 +76,7 @@ furo = "^2023.9.10"
 ipykernel = "^6.26.0"
 ipython = "^8.17.2"
 ipywidgets = "^8.1.1"
-isort = "^5.12.0"
+isort = "^5.13.2"
 mypy = "^1.6.1"
 nbconvert = "^7.10.0"
 nbsphinx = "^0.9.3"
