### 2023-12-19 a133ac0
Bump matplotlib from 3.8.0 to 3.8.2
