+++ pyproject.toml
@@ -50,7 +50,7 @@ Changelog = "https://github.com/robbievanleeuwen/section-properties/releases"
 python = ">=3.9.0,<3.12"
 numpy = "^1.24.0"
 scipy = "^1.10.0"
-matplotlib = "^3.8.0"
+matplotlib = "^3.8.2"
 shapely = "^2.0.2"
 triangle = "^20230923"
 rich = "^13.5.0"
