### 2023-12-19 342e40a
Bump mypy from 1.6.1 to 1.7.1
