+++ pyproject.toml
@@ -77,7 +77,7 @@ ipykernel = "^6.26.0"
 ipython = "^8.17.2"
 ipywidgets = "^8.1.1"
 isort = "^5.12.0"
-mypy = "^1.6.1"
+mypy = "^1.7.1"
 nbconvert = "^7.10.0"
 nbsphinx = "^0.9.3"
 notebook = "^7.0.6"
