+++ docs/requirements.txt
@@ -2,7 +2,7 @@ furo==2023.9.10
 ipykernel==6.26.0
 ipython==8.17.2
 ipywidgets==8.1.1
-nbconvert==7.10.0
+nbconvert==7.13.0
 nbsphinx==0.9.3
 sphinx==7.2.6
 sphinx-click==5.0.1
