+++ pyproject.toml
@@ -78,7 +78,7 @@ ipython = "^8.17.2"
 ipywidgets = "^8.1.1"
 isort = "^5.12.0"
 mypy = "^1.6.1"
-nbconvert = "^7.10.0"
+nbconvert = "^7.13.0"
 nbsphinx = "^0.9.3"
 notebook = "^7.0.6"
 pep8-naming = "^0.13.3"
