### 2023-12-19 5895253
Bump pre-commit from 3.5.0 to 3.6.0
