+++ pyproject.toml
@@ -82,7 +82,7 @@ nbconvert = "^7.10.0"
 nbsphinx = "^0.9.3"
 notebook = "^7.0.6"
 pep8-naming = "^0.13.3"
-pre-commit = "^3.5.0"
+pre-commit = "^3.6.0"
 pre-commit-hooks = "^4.5.0"
 Pygments = "^2.16.1"
 pytest = "^7.4.3"
