### 2023-12-19 454867b
Bump rich from 13.5.0 to 13.7.0
