+++ pyproject.toml
@@ -53,7 +53,7 @@ scipy = "^1.10.0"
 matplotlib = "^3.8.0"
 shapely = "^2.0.2"
 triangle = "^20230923"
-rich = "^13.5.0"
+rich = "^13.7.0"
 click = "^8.1.7"
 more-itertools = "^10.0.0"
 numba = { version = "^0.58.0", optional = true }
