### 2023-12-19 f70fd35
Bump sphinx-click from 5.0.1 to 5.1.0
