+++ docs/requirements.txt
@@ -5,6 +5,6 @@ ipywidgets==8.1.1
 nbconvert==7.10.0
 nbsphinx==0.9.3
 sphinx==7.2.6
-sphinx-click==5.0.1
+sphinx-click==5.1.0
 sphinx-copybutton==0.5.2
 sphinxext-opengraph==0.8.2
