+++ pyproject.toml
@@ -91,7 +91,7 @@ pytest-check = "^2.2.2"
 pyupgrade = "^3.15.0"
 sphinx = "^7.2.6"
 sphinx-autobuild = "^2021.3.14"
-sphinx-click = "^5.0.1"
+sphinx-click = "^5.1.0"
 sphinx-copybutton = "^0.5.2"
 sphinxext-opengraph = "^0.8.2"
 
