### 2023-12-21 #6105 @pprachas
Hi @robbievanleeuwen,

First of all -- Happy Holidays!

I also really appreciate the work that is put into this package -- it looks like a lot of effort is put into the documentation!

Here are a few questions that I have:

1. Installation of both the stable and developer's version to worked seamlessly for me. I might have missed this but is there a way to quickly check if the installation and functionality of the installation of the stable version (e.g. something similar to pytest benchmarks in developer's version, or a simple python script to run). A quick instruction on how to do so will also be helpful in the installation section of the Github page and documentation.

2. Can you actually run the jupyter notebook examples in the documentation? If this is the case, a note on this should also be in the documentation and instructions to optionally install jupyter notebooks should be in the stable installation of the software.

Looking forward to your reply!
