### 2023-12-23 #6105 @robbievanleeuwen
Hi @pprachas, happy holidays to you as well! Thank you very much for taking the time to review `sectionproperties`!

See below response to your questions:

1. As far as I understand, it is not super common to ship the tests with the python package - the [python packaging template](https://github.com/cjolowicz/cookiecutter-hypermodern-python) I used for this package does not do this by default. At the moment, the only way to "verify" the installation would be to run `sectionproperties --version` within the terminal, however this simply prints the installed version. If you think it is worth having an option for users installing via `pip install sectionproperties` to run the tests suite on their machine I can look into adding this functionality 👍.
2. The jupyter notebook examples are run by the CI each time the documentation is built. At the moment, the only way for users to run these examples would be to download the desired example from the [repo](https://github.com/robbievanleeuwen/section-properties/tree/master/docs/examples) and run locally. I could add a quick description of this to the top of [this](https://sectionproperties.readthedocs.io/en/stable/examples.html) page if you think it would be helpful?

Thanks for your comments thus far, please feel free to let me know if you have any more questions!
