### 2024-01-05 #6105 @pprachas
Hi @robbievanleeuwen,

Having instructions to download and run the notebooks will be nice! This will also serve as a way to "verify" the installations (addressing the first point) as well as allow users to test software functionality as well as serve as tutorials.
