### 2024-01-10 #6105 @robbievanleeuwen
Hi @pprachas, happy new year!

Thanks for your comment, I've update the examples page (see [here](https://sectionproperties.readthedocs.io/en/latest/examples.html) in the 'latest' version) to hopefully include the above!
