+++ .github/workflows/tests.yml
@@ -158,4 +158,4 @@ jobs:
           nox --session=coverage -- xml -i
 
       - name: Upload coverage report
-        uses: codecov/codecov-action@v3.1.4
+        uses: codecov/codecov-action@v4.0.1
