### 2024-03-03 #6105 @pprachas
Hi @prashjha and @robbievanleeuwen ,

Sorry for the delay, but I have finished the review. ``sectionproperties``  is extremely handy software that will aid analysis of beams. I am able to install the software as well as reproduce all the examples. Validation examples are compared with known solutions and have acceptable error tolerances. Documentation is comprehensive and well-written. I definitely recommend for ``sectionproperties`` to be accepted in JOSS!
