### 2024-03-08 9564c58
Bump poetry from 1.7.1 to 1.8.2 in /.github/workflows
Bumps [poetry](https://github.com/python-poetry/poetry) from 1.7.1 to 1.8.2.
- [Release notes](https://github.com/python-poetry/poetry/releases)
- [Changelog](https://github.com/python-poetry/poetry/blob/main/CHANGELOG.md)
- [Commits](https://github.com/python-poetry/poetry/compare/1.7.1...1.8.2)

---
updated-dependencies:
- dependency-name: poetry
  dependency-type: direct:production
  update-type: version-update:semver-minor
...

Signed-off-by: dependabot[bot] <support@github.com>
