+++ .github/workflows/constraints.txt
@@ -1,5 +1,5 @@
 pip==24.0
 nox==2023.4.22
 nox-poetry==1.0.3
-poetry==1.7.1
+poetry==1.8.2
 virtualenv==20.25.1
