+++ src/sectionproperties/analysis/__init__.py
@@ -1,2 +1,3 @@
 """sectionproperties analysis module."""
+
 from sectionproperties.analysis.section import Section
