+++ docs/requirements.txt
@@ -1,10 +1,10 @@
-furo==2023.9.10
-ipykernel==6.27.1
+furo==2024.01.29
+ipykernel==6.29.3
 ipython==8.18.1
-ipywidgets==8.1.1
-nbconvert==7.13.0
+ipywidgets==8.1.2
+nbconvert==7.16.2
 nbsphinx==0.9.3
 sphinx==7.2.6
 sphinx-click==5.1.0
 sphinx-copybutton==0.5.2
-sphinxext-opengraph==0.8.2
+sphinxext-opengraph==0.9.1
