+++ pyproject.toml
@@ -65,36 +65,36 @@ intel-openmp = { version = "==2023.2.0", optional = true }
 mkl = { version = "==2023.2.0", optional = true }
 
 [tool.poetry.group.dev.dependencies]
-black = "^23.10.1"
-coverage = { extras = ["toml"], version = "^7.3.3" }
+black = "^24.2.0"
+coverage = { extras = ["toml"], version = "^7.4.3" }
 darglint = "^1.8.1"
-flake8 = "^6.1.0"
-flake8-bugbear = "^23.12.2"
+flake8 = "^7.0.0"
+flake8-bugbear = "^24.2.6"
 flake8-docstrings = "^1.7.0"
 flake8-pytest-style = "^1.7.2"
 flake8-rst-docstrings = "^0.3.0"
-furo = "^2023.9.10"
-ipykernel = "^6.27.1"
-ipython = "^8.18.1"
-ipywidgets = "^8.1.1"
+furo = "^2024.01.29"
+ipykernel = "^6.29.3"
+ipython = "^8.18.1"  # 8.19 and above drops 3.9 support
+ipywidgets = "^8.1.2"
 isort = "^5.13.2"
-mypy = "^1.7.1"
-nbconvert = "^7.13.0"
+mypy = "^1.9.0"
+nbconvert = "^7.16.2"
 nbsphinx = "^0.9.3"
-notebook = "^7.0.6"
+notebook = "^7.1.1"
 pep8-naming = "^0.13.3"
-pre-commit = "^3.6.0"
+pre-commit = "^3.6.2"
 pre-commit-hooks = "^4.5.0"
 Pygments = "^2.17.2"
-pytest = "^7.4.3"
+pytest = "^8.0.2"
 pytest-benchmark = { extras = ["histogram"], version = "^4.0.0" }
-pytest-check = "^2.2.2"
-pyupgrade = "^3.15.0"
+pytest-check = "^2.3.1"
+pyupgrade = "^3.15.1"
 sphinx = "^7.2.6"
-sphinx-autobuild = "^2021.3.14"
+sphinx-autobuild = "^2024.2.4"
 sphinx-click = "^5.1.0"
 sphinx-copybutton = "^0.5.2"
-sphinxext-opengraph = "^0.8.2"
+sphinxext-opengraph = "^0.9.1"
 
 [tool.poetry.extras]
 dxf = ["cad-to-shapely"]
