### 2024-03-09 b87519d
Update core dependencies
