+++ pyproject.toml
@@ -48,19 +48,19 @@ Changelog = "https://github.com/robbievanleeuwen/section-properties/releases"
 
 [tool.poetry.dependencies]
 python = ">=3.9.0,<3.12"
-numpy = "^1.24.0"
-scipy = "^1.10.0"
-matplotlib = "^3.8.2"
-shapely = "^2.0.2"
+numpy = "^1.24.4"
+scipy = "^1.10.1"
+matplotlib = "^3.8.3"
+shapely = "^2.0.3"
 triangle = "^20230923"
-rich = "^13.7.0"
+rich = "^13.7.1"
 click = "^8.1.7"
-more-itertools = "^10.0.0"
-numba = { version = "^0.58.0", optional = true }
+more-itertools = "^10.2.0"
+numba = { version = "^0.59.0", optional = true }
 cad-to-shapely = { version = "^0.3.1", optional = true }
 rhino-shapley-interop = { version = "^0.0.4", optional = true }
-rhino3dm = { version = "==8.0.0b3", optional = true }
-pypardiso = { version = "^0.4.4", optional = true }
+rhino3dm = { version = "^8.4.0", optional = true }
+pypardiso = { version = "^0.4.5", optional = true }
 intel-openmp = { version = "==2023.2.0", optional = true }
 mkl = { version = "==2023.2.0", optional = true }
 
