+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "sectionproperties"
-version = "3.1.3"
+version = "3.2.0"
 description = "A python package for the analysis of arbitrary cross-sections using the finite element method."
 license = "MIT"
 authors = [
