### 2024-03-12 #6105 @dbrizard
Hi @robbievanleeuwen
I have completed the review.
`sectionproperties` is a mature project with a lot of functionalities.
The documentation is well organized, very detailed and full of examples and bits of code.
I have also appreciated the Theory section of the documentation.
Installation was really easy, which is noticeable.
All the bits of code I copied worked seamlessly.

I really recommend publishing `sectionproperties` in JOSS.
