### 2024-04-15 #6105 @robbievanleeuwen
Hi @prashjha thanks for the update. The project has been updated to `v3.2.1` and this latest version has been archived at https://zenodo.org/records/10971849.
