### 2024-04-17 #6105 @prashjha
@robbievanleeuwen, it seems there is a problem with one of the references; see [here](https://github.com/openjournals/joss-reviews/issues/6105#issuecomment-2060982934). Please can you fix it?
