### 2024-04-17 #6105 @kyleniemeyer
@robbievanleeuwen in addition, our paper builder is having trouble with the `\rm` command you used in a few places—can you replace this with `\textrm{}`? The issue might also be the missing brackets around the variable this is applied to.

I also think it is having an issue with the equation where you use `\bmatrix`, since it is missing `&` between the two columns.
