### 2024-04-19 #6105 @robbievanleeuwen
Hi @kyleniemeyer & @prashjha, I changed four occurences of `rm` to `textrm` with the latest commit to the `master` branch (robbievanleeuwen/section-properties@e01f3ee), one of which was embedded in a `bmatrix`. Hopefully this resolves the above four errors!
