### 2024-04-21 #6105 @robbievanleeuwen
Many thanks to @dbrizard and @pprachas for their assistance in reviewing `sectionproperties`, and to @prashjha and @kyleniemeyer for editing! We appreciate you volunteering your time to this great cause :) Also tagging my co-author @connorferster in case you missed this!
