### 2024-04-21 #6105 @robbievanleeuwen
Also @kyleniemeyer, not sure if this is an easy fix, but my name is incorrectly displayed on the [JOSS website](https://joss.theoj.org/papers/10.21105/joss.06105) and resulting citation. It reads correctly in the pdf paper (Robbie van Leeuwen), however displays "Robbie Leeuwen" elsewhere.
