### 2024-04-21 #6105 @kyleniemeyer
@robbievanleeuwen ah, good question. Fixable yes, but might require a manual fix.

@openjournals/dev how can we update the author's name on the JOSS website, currently displayed as "Robbie Leeuwen" but should be "Robbie van Leeuwen"? The paper itself is correct.
